+++ library/Kathu/World/Field.hs
@@ -7,16 +7,16 @@ module Kathu.World.Field where
 import Control.Lens
 import Control.Monad
 import Control.Monad.IO.Class (MonadIO, liftIO)
-import Data.Map (Map)
-import qualified Data.Map as Map
+import Control.Monad.ST (RealWorld)
+import Data.Map.Strict (Map)
+import qualified Data.Map.Strict as Map
 import Data.Maybe (maybe)
-import Data.Vector (Vector)
 import qualified Data.Vector as Vec
-import Data.Vector.Mutable (IOVector)
-import qualified Data.Vector.Mutable as MVec
+import qualified Data.Vector.Unboxed.Mutable as UMVec
+import Linear.V3 (V3(..))
+
 import Kathu.Util.MultiDimVector
 import Kathu.World.Tile
-import Linear.V3 (V3(..))
 
 unitsPerTile :: Num a => a
 unitsPerTile = 16
@@ -29,16 +29,16 @@ fieldSize = 32
 fieldHeight :: Num a => a
 fieldHeight = 4
 
-type FieldSet g = Map (V3 Int) (Field g)
+type FieldSet = Map (V3 Int) Field
 
-newtype Field g = Field
-    { fieldData :: (IOVector (TileState g))
+newtype Field = Field
+    { fieldData :: MVector RealWorld (TileState)
     --, fieldIndRenders :: (IOVector Bool)
     --, fieldPreRenders :: [Image]
     }
 
-mkField :: MonadIO m => m (Field g)
-mkField = liftIO . fmap Field . MVec.replicate size . mkTileState $ emptyTile
+mkField :: MonadIO m => m (Field)
+mkField = liftIO . fmap Field . UMVec.replicate size . mkTileState $ emptyTile
     where size = fieldSize * fieldSize * fieldHeight
 
 indexFromCoord :: V3 Int -> Int
@@ -52,34 +52,33 @@ worldCoordFromTile :: Num a => V3 Int -> V3 Int -> V3 a
 worldCoordFromTile (V3 fx fy fz) (V3 lx ly lz) = V3 (conv fieldSize fx lx) (conv fieldSize fy ly) (conv fieldHeight fz lz)
     where conv s f l = unitsPerTile * fromIntegral ((f * s) + l)
 
-fetchTileState :: MonadIO m => V3 Int -> Field g -> m (TileState g)
-fetchTileState v (Field tilest) = liftIO $ MVec.read tilest (indexFromCoord v)
+fetchTileState :: MonadIO m => V3 Int -> Field -> m (TileState)
+fetchTileState v (Field tilest) = liftIO $ UMVec.read tilest (indexFromCoord v)
 
-setTileState :: MonadIO m => V3 Int -> TileState g -> Field g -> m ()
-setTileState v t (Field tilest) = liftIO $ MVec.write tilest (indexFromCoord v) t
+setTileState :: MonadIO m => V3 Int -> TileState -> Field -> m ()
+setTileState v t (Field tilest) = liftIO $ UMVec.write tilest (indexFromCoord v) t
 
-foreachTile :: MonadIO m => (TileState g -> m a) -> Field g -> m ()
+foreachTile :: MonadIO m => (TileState -> m a) -> Field -> m ()
 foreachTile f (Field tilest) = go 0
-    where len = MVec.length tilest
+    where len = UMVec.length tilest
           go i | i == len  = pure ()
-               | otherwise = liftIO (MVec.read tilest i) >>= f >> go (i + 1)
+               | otherwise = liftIO (UMVec.read tilest i) >>= f >> go (i + 1)
 
-fieldFoldM :: MonadIO m => (a -> V3 Int -> TileState g -> m a) -> a -> Field g -> m a
+fieldFoldM :: MonadIO m => (a -> V3 Int -> TileState -> m a) -> a -> Field -> m a
 fieldFoldM f !acc (Field tiles) = go 0 0 0 acc
     where go !x !y !z !b | z == fieldHeight = pure b
                          | y == fieldSize   = go 0 0 (z + 1) b
                          | x == fieldSize   = go 0 (y + 1) z b
-                         | otherwise        = let v = V3 x y z in liftIO (MVec.read tiles (indexFromCoord v)) >>= f b v >>= go (x + 1) y z
+                         | otherwise        = let v = V3 x y z in liftIO (UMVec.read tiles (indexFromCoord v)) >>= f b v >>= go (x + 1) y z
 
 -- Conversion
 
-mkFields :: MonadIO m => Int -> Int -> Int -> m (Vector3D (Field g))
+mkFields :: MonadIO m => Int -> Int -> Int -> m (Vector3D (Field))
 mkFields x y z = repliVecM x (repliVecM y (repliVecM z mkField))
-    where repliVecM :: MonadIO m => Int -> m a -> m (Vector a)
-          repliVecM n = fmap Vec.fromList . replicateM n
+    where repliVecM n = fmap Vec.fromList . replicateM n
    
-fromTileList :: MonadIO m => Vector3D (Tile g) -> m (FieldSet g)
-fromTileList tiles = if zLayers == 0 || yLayers == 0 || xLayers == 0 then empty else buildNew 
+fromTileVector3D :: MonadIO m => Vector3D (Tile g) -> m FieldSet
+fromTileVector3D tiles = if zLayers == 0 || yLayers == 0 || xLayers == 0 then empty else buildNew 
     where empty                       = Map.singleton (V3 0 0 0) <$> mkField
           minFields :: (Integral a) => Float -> a -> Int
           minFields s                 = ceiling . (/s) . fromIntegral
@@ -93,10 +92,10 @@ fromTileList tiles = if zLayers == 0 || yLayers == 0 || xLayers == 0 then empty 
                      | x == xLen = go 0 (y + 1) z v
                      | otherwise = (runForTile (read3D z y x tiles) writeTile) >> go (x + 1) y z v
               where (V3 fx fy fz) = getContField x y z
-                    runForTile t f | t ^. tileID == emptyTileID = pure () -- do nothing if wanting to right empty tile
-                                   | otherwise = liftIO (mRead3D fx fy fz v) >>= \case
+                    runForTile t f | t^.tileID == emptyTileID = pure () -- do nothing if wanting to write empty tile
+                                   | otherwise                = liftIO (mRead3D fx fy fz v) >>= \case
                                        Just (m) -> f t m
                                        Nothing  -> mkField >>= \field -> liftIO (mWrite3D fx fy fz (Just field) v) >> f t field
-                    writeTile t field = setTileState (V3 x y z) (mkTileState t) field
+                    writeTile t field = mkTileStateWithMetadata t >>= \newTileState -> setTileState (V3 x y z) newTileState field
           convertToMap iovec = fmap Map.fromList . liftIO .  miFoldl3D foldf [] $ iovec
               where foldf acc x y z cur = pure $ maybe acc ((:acc) . (V3 x y z,)) cur